requirements.txt
# Python dependencies for Patient Discharge Automation System

# LangChain and LangGraph
langchain==0.3.13
langchain-core==0.3.28
langchain-google-genai==2.0.8
langgraph==0.2.59

# Google Generative AI
google-generativeai==0.8.3

# Data validation
pydantic==2.10.4
pydantic-core==2.27.2

# Environment management
python-dotenv==1.0.1

# Utilities
typing-extensions==4.12.2
